First pass creating a chirper class to wrap the twitter gem
It's better to not code the app against the 3rd party lib, instead we
hide it behind our own interface which will allow us to swap it out if
we need to. Plus we only need to implement enough interface for our
needs.

Additionally we also want to remove the global $client variable